Casts unsigned to signed to make negative.

+++ src/Graphics/Quad.cpp
@@ -101,8 +101,9 @@ void zQuad::LoadRegularVertices()
 void zQuad::LoadInvertedVertices()
 {
     m_DrawData.Vertices[0].position = math::vector_t(0, 0);
-    m_DrawData.Vertices[1].position = math::vector_t(0, -m_BoundingBox.h);
-    m_DrawData.Vertices[2].position = math::vector_t(m_BoundingBox.w, -m_BoundingBox.h);
+    m_DrawData.Vertices[1].position = math::vector_t(0, -static_cast<int>(m_BoundingBox.h));
+    m_DrawData.Vertices[2].position = math::vector_t(m_BoundingBox.w,
+                                   -static_cast<int>(m_BoundingBox.h));
     m_DrawData.Vertices[3].position = math::vector_t(m_BoundingBox.w, 0);
 }
